#68 Separate Responsibilities for Magento Provider
